pptx-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
httpx==0.28.1
python-pptx==1.0.2
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.19
